Add a friendly check for the font-awesome bower path
I use this in a couple of my addons to insure the required Bower
packages are installed. Simply checks the file system for the
font-awesome path. If not available, a friendly error message is
presented informing the user how to fix the problem. The
`target.import()` later on would have failed anyway, so this is just an
up-front check.