Adjust ember-cli-sass options to include font-awesome path
When the `useScss` option is enabled, adjust the `ember-cli-sass`
options to include the font-awesome path. That way the user can
`@import “font-awesome”;` instead of needing to include the bower path.
Also make sure this addon is ran before the `ember-cli-sass` addon, so
our option changes take effect. The README is also updated to better
explain Sass support and show this new usage.