Adjust ember-cli-less options to include font-awesome path
Similar to the Sass support, when the `useLess` option is enabled,
adjust the `ember-cli-less` options to include the font-awesome path.
That way the user can just `@import “font-awesome”;` in `app.less`.
Also make sure this addon is ran before the `ember-cli-less` addon, so
our option changes take effect. The README is also updated to add this
new option and explain its usage.