Check for fonts already imported
Added logic to check for font files that already have been imported
into the fonts folder, and skip the import if that’s the case (avoiding
an ember-cli error). Also notify the user which files were skipped and
how to fix the duplication, but continue with the build since
everything is technically Ok. This applies to issue #84